refactor(all): 폴더구조를 MVC 패턴에 맞게 분리
