+++ src/main/java/baseball/domain/GameResult.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 import java.util.List;
