+++ src/main/java/baseball/domain/RestartDecision.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.domain;
 
 public enum RestartDecision {
 	RESTART(1),
